+++ nn/parameters.py
@@ -85,46 +85,46 @@ class Parameters:
     def initiate_zeros(self,in_dim,out_dim,layer_num):
         self.__check_attributes(layer_num,in_dim)
         self.__parameters_num += out_dim * (in_dim+1)
-        W = np.zeros((out_dim,in_dim), dtype=np.double)
-        b = np.zeros((out_dim,1), dtype=np.double)
+        W = np.zeros((out_dim,in_dim))
+        b = np.zeros((out_dim,1))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_random(self,in_dim,out_dim,layer_num):
         self.__check_attributes(layer_num,in_dim)
         self.__parameters_num += out_dim * (in_dim + 1)
         W = np.random.randn(out_dim,in_dim)
-        b = np.zeros((out_dim,1), dtype=np.double)
+        b = np.zeros((out_dim,1))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_xavier(self,in_dim,out_dim,layer_num):
         self.__check_attributes(layer_num,in_dim)
         self.__parameters_num += out_dim * (in_dim + 1)
         W = np.sqrt(1 / in_dim) * np.random.randn(out_dim, in_dim) 
-        b = np.zeros((out_dim, 1), dtype=np.double)
+        b = np.zeros((out_dim, 1))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_zeros_conv2(self,f1, f2, n_C_prev, n_C,layer_num):
         self.__parameters_num += (f1*f2*n_C_prev+1) * (n_C)
-        W = np.zeros((f1, f2, n_C_prev, n_C), dtype=np.double)
-        b = np.zeros((1,1,1,n_C), dtype=np.double)
+        W = np.zeros((f1, f2, n_C_prev, n_C))
+        b = np.zeros((1,1,1,n_C))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_random_conv2(self, f1, f2, n_C_prev, n_C,layer_num):
         self.__parameters_num += (f1*f2*n_C_prev+1) * (n_C)
         W = np.random.randn(f1, f2, n_C_prev, n_C)
-        b = np.zeros((1,1,1,n_C), dtype=np.double)
+        b = np.zeros((1,1,1,n_C))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_xavier_conv2(self,f1, f2, n_C_prev, n_C,layer_num):
         self.__parameters_num += (f1*f2*n_C_prev+1) * (n_C)
         W = np.sqrt(1 / n_C_prev) * np.random.randn(f1, f2, n_C_prev, n_C)
-        b = np.zeros((1,1,1,n_C), dtype=np.double)
+        b = np.zeros((1,1,1,n_C))
         self.__add_weights(W,b,'W','b',layer_num)
 
     def initiate_batchnorm_params(self, n_channels, layer_num):
         self.__parameters_num += 4*n_channels
-        gamma = np.ones((1,1,n_channels,1), dtype=np.double)
-        beta = np.zeros((1,1,n_channels,1), dtype=np.double)
+        gamma = np.ones((1,1,n_channels,1))
+        beta = np.zeros((1,1,n_channels,1))
         self.__add_weights(gamma,beta,'gamma','beta',layer_num)
 
     def get_batchnorm_params(self, layer_num):
